remove /api subdirectory
