+++ CHANGES.txt
@@ -7,6 +7,7 @@
 - Remove titiler.core in favor of starlette settings (#55, author @geospatial-jeff)
 - Add fastapi exception handlers (#56, author @geospatial-jeff)
 - Remove intermediary routers (#57, author @geospatial-jeff)
+- Remove /titiler/api submodule (e.g titiler.api.utils -> titiler.utils)
 
 2.1.2 (2020-06-24)
 ------------------
