--- titiler/api/__init__.py
@@ -1,1 +1,0 @@
-"""titiler.api"""
