+++ tox.ini
@@ -3,7 +3,7 @@ envlist = py37
 
 
 [testenv]
-extras = test,stac,mosaic
+extras = test,mosaic
 commands=
     python -m pytest --cov titiler --cov-report term-missing --ignore=venv
 deps=
